Add "principal epochs" to allow safeish reuse of a principal ID.
Specific *instances* of a principal ID are named with a random epoch
identifier. When querying, it can be omitted, meaning "any or all
relevant epochs". This makes it possible to talk about a principal ID
without having to specify an epoch, allowing e.g. "password reset" to
take effect without having to wipe/reconfigure server machines. On the
other hand, one may specify a specific epoch, meaning that those
machines will not be affected by an "account reset": if someone
compromises the email address, all they can do is set up a new epoch,
and so those machines will not see it; a kind of fail safe.

This is working toward a sensible signup flow.